Remove keyboard specific stuff for now
So this patch removes the keyboard specific stuff and resolves the weirdness I was seeing. I've also tested this on a laptop with the original package and this mod to confirm nothing has been broken. From what I can tell nothing was affected there by this change.

(Did learn though that depending on the laptop scroll perspective may make the settings confusing.)

Another important thing to note is that it's common for shift scroll to be horizontal scroll instead of vertical. Lots of browsers will do this and from my testing my laptop did this in atom even (desktop does not).

So while this does work as is, I'd say this should be addressed differently in the long run. Not sure what that should look like though.